LPS-188721 Extract service from PortletPreferencesModelListener to PortletPreferencesConfigurationListener
